spec: allow node publish volume in detached mode
Detached mode is a special mode that a volume is not present
in any host path (thus ignoring `target_path` field). This can be
useful if a runtime can handle the volume directly. Hypervisor
based runtimes like kata can benifit from this when dealing with
remote storages like ceph rbd, NFS and cifs.

Signed-off-by: Peng Tao <bergwolf@gmail.com>